Move components into page-specific directories
