+++ src/components/general/notice_settings.tsx
