+++ src/pages/users/setting/general.tsx
@@ -4,10 +4,10 @@ import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
 import { ILocale } from '@/interfaces/locale';
 import Head from 'next/head';
 import Layout from '@/components/beta/layout/layout';
-import UserSettings from '@/components/user_settings/user_settings';
-import NoticeSettings from '@/components/notice_settings/notice_settings';
-import AccountBookSettings from '@/components/account_book_settings/account_book_settings';
-import AccountSettings from '@/components/account_settings/account_settings';
+import UserSettings from '@/components/general/user_settings/user_settings';
+import NoticeSettings from '@/components/general/notice_settings';
+import AccountBookSettings from '@/components/general/account_book_settings/account_book_settings';
+import AccountSettings from '@/components/general/account_settings/account_settings';
 import APIHandler from '@/lib/utils/api_handler';
 import { IUserSetting } from '@/interfaces/user_setting';
 import { APIName } from '@/constants/api_connection';
